requirements.txt
certifi==2017.11.5
chardet==3.0.4
click==6.7
dominate==2.3.1
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-MySQLdb==0.2.0
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
mysqlclient==1.3.12
pkg-resources==0.0.0
PyMySQL==0.7.11
python-dotenv==0.7.1
requests==2.18.4
urllib3==1.22
visitor==0.1.3
Werkzeug==0.12.2
